plugins/tools-document/skills/pptx/scripts/requirements.txt
defusedxml>=0.7.0
lxml>=4.9.0
Pillow>=9.0.0
